requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
astroid==3.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
docopt==0.6.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
mccabe==0.7.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
pip-autoremove==0.10.0
pip-tools==7.3.0
pipreqs==0.4.13
platformdirs==3.11.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==3.0.2
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
sympy==1.12
termcolor==2.3.0
tiktoken==0.5.1
tokenizers==0.14.1
tomlkit==0.12.1
torchaudio==2.1.0
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
yarg==0.1.9
yarl==1.9.2
